Criando validações dos produtos

+++ src/services/products/DeleteProductService.ts
@@ -2,6 +2,7 @@ import Product from "../../models/products/Produto";
 
 
 export const DeleteProductService = async ({ id_produto }: { id_produto: number }) => {
+    if (!id_produto) return {error: 'Id produto é requerido!'  }
 
     const product = await Product.destroy({
         where: {
